requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless>=4.8.0
pillow>=9.5.0
numpy>=1.24.0
pyyaml>=6.0
requests>=2.31.0
huggingface-hub>=0.23.0
